fix: improve authentication & refactor code
