+++ agents/langgraph/templates/agentic_rag/data/load_documents_ai4rag.py
@@ -8,8 +8,12 @@ NOTE: Use load_documents_wrapper.py as entry point to fix sqlite3 issue.
 """
 
 import os
+import sys
 import uuid
 from datetime import datetime
+from pathlib import Path
+
+sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 
 from ai4rag.rag.embedding.openai_model import (
     OpenAIEmbeddingModel,
@@ -19,6 +23,7 @@ from ai4rag.rag.vector_store import get_vector_store, get_vector_store_config
 from dotenv import load_dotenv
 from langchain_community.document_loaders import TextLoader
 from langchain_text_splitters import RecursiveCharacterTextSplitter
+from milvus_cert_helper import normalize_milvus_cert
 from openai import OpenAI
 
 load_dotenv(verbose=True)
@@ -50,9 +55,6 @@ def load_and_index_documents(
     # Get Milvus configuration
     milvus_uri = os.getenv("MILVUS_URI")
     milvus_token = os.getenv("MILVUS_TOKEN")
-    milvus_cert = os.getenv(
-        "MILVUS_SERVER_CERT"
-    )  # None if not set, empty string if set to ""
 
     # Generate a unique collection name with timestamp if not provided
     milvus_collection = os.getenv("MILVUS_COLLECTION_NAME")
@@ -67,25 +69,8 @@ def load_and_index_documents(
     if not milvus_uri:
         raise ValueError("MILVUS_URI must be set")
 
-    # Handle MILVUS_SERVER_CERT - can be PEM text or path to file
-    # If it looks like a file path (doesn't start with -----BEGIN), read it
-    # Otherwise, assume it's already PEM text
-    if milvus_cert and not milvus_cert.startswith("-----BEGIN"):
-        # It's a file path - read the certificate content
-        if not os.path.exists(milvus_cert):
-            raise ValueError(f"MILVUS_SERVER_CERT file not found at {milvus_cert}")
-        with open(milvus_cert, "r") as f:
-            cert_content = f.read()
-        os.environ["MILVUS_SERVER_CERT"] = cert_content
-        print(f"✓ Loaded certificate from {milvus_cert}")
-    elif milvus_cert and milvus_cert.startswith("-----BEGIN"):
-        # It's already PEM text - use it directly
-        os.environ["MILVUS_SERVER_CERT"] = milvus_cert
-        print("✓ Using certificate from environment variable (PEM text)")
-    else:
-        print(
-            "⚠ No MILVUS_SERVER_CERT provided - connection may fail if TLS is required"
-        )
+    # Normalize MILVUS_SERVER_CERT (file path → PEM text)
+    normalize_milvus_cert()
 
     # Set other environment variables for ai4rag
     os.environ["MILVUS_URI"] = milvus_uri
