+++ agents/langgraph/templates/agentic_rag/main.py
@@ -1,41 +1,26 @@
 import asyncio
 import json
 import logging
-import sys
 import time
 import uuid
 from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
 from os import getenv
-from pathlib import Path
 from typing import Any
 
-# Fix sqlite3 version issue for chromadb (required by ai4rag)
-# Must be done BEFORE importing agentic_rag
-try:
-    import pysqlite3  # noqa: F401
+from sqlite_shim import patch_sqlite3
 
-    sys.modules["sqlite3"] = sys.modules.pop("pysqlite3")
-except ImportError:
-    pass  # pysqlite3-binary not available, continue with system sqlite3
+patch_sqlite3()
 
-import openai
-import requests as http_requests
-from agentic_rag.agent import get_graph_closure
-from agentic_rag.tracing import enable_tracing
-from fastapi import FastAPI, HTTPException
-from fastapi.openapi.docs import get_swagger_ui_html
-from fastapi.responses import (
-    FileResponse,
-    HTMLResponse,
-    JSONResponse,
-    RedirectResponse,
-    StreamingResponse,
-)
-from langchain_core.exceptions import OutputParserException
-from langchain_core.messages import AIMessage, HumanMessage, ToolMessage
-from langgraph.errors import GraphRecursionError
-from pydantic import BaseModel, Field, ValidationError
+import openai  # noqa: E402
+from agentic_rag.agent import get_graph_closure  # noqa: E402
+from agentic_rag.tracing import enable_tracing  # noqa: E402
+from fastapi import FastAPI, HTTPException  # noqa: E402
+from fastapi.responses import JSONResponse, StreamingResponse  # noqa: E402
+from langchain_core.exceptions import OutputParserException  # noqa: E402
+from langchain_core.messages import AIMessage, HumanMessage, ToolMessage  # noqa: E402
+from langgraph.errors import GraphRecursionError  # noqa: E402
+from pydantic import BaseModel, Field, ValidationError  # noqa: E402
 
 logger = logging.getLogger(__name__)
 
@@ -184,75 +169,12 @@ app = FastAPI(
         "To access the sandbox playground, click "
         "[Sandbox Playground](/playground)."
     ),
-    docs_url=None,
     lifespan=lifespan,
     openapi_tags=[
         {"name": "Health", "description": "Service health monitoring"},
         {"name": "Chat", "description": "Chat completion operations"},
     ],
 )
-
-
-_PLAYGROUND_URL = getenv("PLAYGROUND_URL", "http://localhost:5002").rstrip("/")
-_SANDBOX_MODE = bool(getenv("K8S_REVIEWER_TOKEN", "").strip())
-_PLAYGROUND_TOKEN = getenv("PLAYGROUND_TOKEN", "").strip()
-
-
-@app.get(
-    "/docs",
-    include_in_schema=False,
-    response_class=HTMLResponse,
-)
-async def custom_swagger_ui():
-    """Serve Swagger UI with a shortcut to the sandbox playground."""
-    swagger_html = get_swagger_ui_html(
-        openapi_url=app.openapi_url,
-        title=f"{app.title} - Swagger UI",
-        swagger_favicon_url="https://fastapi.tiangolo.com/img/favicon.png",
-    )
-    html = swagger_html.body.decode("utf-8")
-    shortcut = (
-        '<a href="/playground" class="sandbox-playground-link" '
-        'title="Open Sandbox Playground">🎮 Sandbox Playground</a>'
-    )
-    styles = """
-    <style>
-      .sandbox-playground-link {
-        position: fixed;
-        top: 18px;
-        right: 24px;
-        z-index: 10;
-        padding: 8px 12px;
-        border-radius: 4px;
-        color: #fff;
-        background: #3b4151;
-        font: 600 13px sans-serif;
-        text-decoration: none;
-      }
-      .sandbox-playground-link:hover { background: #2f3545; }
-    </style>
-    """
-    html = html.replace("<body>", f"<body>{styles}{shortcut}", 1)
-    return HTMLResponse(html)
-
-
-@app.get(
-    "/playground",
-    summary="Open the sandbox playground",
-    description=(
-        "Redirects to the local sandbox playground UI. Start it with "
-        "`make playground-sandbox`; it obtains the ServiceAccount token "
-        "automatically through `oc`."
-    ),
-    include_in_schema=False,
-    tags=["Playground"],
-    responses={307: {"description": "Redirect to the sandbox playground UI"}},
-)
-async def playground_redirect():
-    """Serve the sandbox playground or redirect to the local fallback UI."""
-    if _SANDBOX_MODE:
-        return FileResponse(_SANDBOX_PLAYGROUND_HTML)
-    return RedirectResponse(url=_PLAYGROUND_URL)
 
 
 def _auth_enabled() -> bool:
@@ -601,93 +523,13 @@ async def health():
     return body
 
 
-# ── Playground UI ────────────────────────────────────────────────────────────
-_BASE_DIR = Path(__file__).resolve().parent
-_PLAYGROUND_HTML = _BASE_DIR / "playground" / "templates" / "index.html"
-_SANDBOX_PLAYGROUND_HTML = _BASE_DIR / "playground-sandbox" / "templates" / "index.html"
-# In Docker the images are copied to /opt/app-root/src/images; locally they live at the repo root
-_IMAGES_DIR = _BASE_DIR / "images"
-if not _IMAGES_DIR.is_dir():
-    _IMAGES_DIR = _BASE_DIR.parent.parent.parent / "images"
+# ── Playground UI (sandbox mode) ─────────────────────────────────────────────
+# Include sandbox playground routes only when K8S_REVIEWER_TOKEN is set
+_SANDBOX_MODE = bool(getenv("K8S_REVIEWER_TOKEN", "").strip())
+if _SANDBOX_MODE:
+    from playground_sandbox import router as sandbox_router
 
-
-@app.get("/", response_class=HTMLResponse, include_in_schema=False)
-async def playground():
-    """Serve the playground chat UI."""
-    if _SANDBOX_MODE:
-        return FileResponse(_SANDBOX_PLAYGROUND_HTML)
-    if _auth_enabled():
-        raise HTTPException(status_code=404, detail="Not found")
-    return FileResponse(_PLAYGROUND_HTML)
-
-
-@app.get("/api/health", include_in_schema=False)
-async def playground_health():
-    """Expose the agent health response expected by the sandbox UI."""
-    return await health()
-
-
-@app.post("/api/chat", include_in_schema=False)
-async def playground_chat(request: ChatCompletionRequest):
-    """Proxy sandbox UI requests with the server-side ServiceAccount token."""
-    if not (_SANDBOX_MODE and _PLAYGROUND_TOKEN):
-        raise HTTPException(
-            status_code=503, detail="Sandbox playground is not configured"
-        )
-
-    payload = request.model_dump(exclude_none=True)
-    payload["stream"] = True
-
-    def event_generator():
-        try:
-            with http_requests.post(
-                "http://127.0.0.1:8080/chat/completions",
-                json=payload,
-                headers={"X-Api-Key": _PLAYGROUND_TOKEN},
-                stream=True,
-                timeout=(10, 300),
-            ) as response:
-                if response.status_code != 200:
-                    error = json.dumps(
-                        {
-                            "error": {
-                                "message": f"Agent returned {response.status_code}: {response.text[:500]}"
-                            }
-                        }
-                    )
-                    yield f"data: {error}\n\n"
-                    return
-                for chunk in response.iter_content(
-                    chunk_size=None, decode_unicode=True
-                ):
-                    if chunk:
-                        yield chunk
-        except http_requests.RequestException as exc:
-            logger.exception("Sandbox playground proxy request failed")
-            error = json.dumps({"error": {"message": str(exc)}})
-            yield f"data: {error}\n\n"
-
-    return StreamingResponse(
-        event_generator(),
-        media_type="text/event-stream",
-        headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
-    )
-
-
-@app.get("/images/{filename:path}", include_in_schema=False)
-async def serve_image(filename: str):
-    """Serve images from the project-level images directory."""
-    if _auth_enabled():
-        raise HTTPException(status_code=404, detail="Not found")
-    base = _IMAGES_DIR.resolve()
-    file_path = (base / filename).resolve()
-    try:
-        file_path.relative_to(base)
-    except ValueError:
-        raise HTTPException(status_code=404, detail="Image not found")
-    if not file_path.is_file():
-        raise HTTPException(status_code=404, detail="Image not found")
-    return FileResponse(file_path)
+    app.include_router(sandbox_router)
 
 
 if __name__ == "__main__":
